requirements.txt
asgiref==3.4.1
astroid==2.4.2
autopep8==1.5.4
Babel==2.9.1
backports.entry-points-selectable==1.1.0
click==8.0.0
distlib==0.3.2
Django==3.2.6
djangorestframework==3.12.4
filelock==3.0.12
Flask==2.0.0
Flask-Moment==0.11.0
Flask-WTF==0.14.3
isort==5.7.0
itsdangerous==2.0.0
Jinja2==3.0.0
lazy-object-proxy==1.4.3
MarkupSafe==2.0.0
mccabe==0.6.1
platformdirs==2.3.0
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.6.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.7.2
Werkzeug==2.0.0
wrapt==1.12.1
WTForms==2.3.3
